requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
aiohttp==3.10.5
google-cloud-bigquery==3.25.0